hotfix

+++ src/grammar.rs
@@ -70,7 +70,7 @@ fn arg<'a>(p: &P<'a>, following: bool) -> PResult<'a> {
         VAR => Ok(p.advance_close(m, Var(p.slice()))),
         KEYWORD => Ok(p.advance_close(m, Keyword(p.slice()))),
         INT => Ok(p.advance_close(m, Int(p.slice().parse().unwrap()))),
-        INT => Ok(p.advance_close(m, Float(p.slice().parse().unwrap()))),
+        FLOAT => Ok(p.advance_close(m, Float(p.slice().parse().unwrap()))),
         STRING => {
             let s = p.slice();
             Ok(p.advance_close(m, Str(&s[1..s.len() - 1])))
